Fix wrong panel link (#18)
Link to /jobs results in 404, the working link is http://okfn.org/about/jobs.
See also: https://github.com/okfn/opendefinition/pull/149

It looks like the opendefinition version of the panel is a copy of this one, so I guess it needs fixed in both places. Please correct this if wrong.